# docs/api/v1-openapi.json
{
  "openapi": "3.0.0",
  "paths": {
...
    "/v1/auth/login": {
      "post": {
        "operationId": "AuthController_login_v1",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/LoginDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "Auth"
        ]
      }
    },
    "/v1/auth/refresh": {
      "post": {
        "operationId": "AuthController_refresh_v1",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/RefreshDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "Auth"
        ]
      }
    },
...
    "/v1/companies/{id}": {
      "get": {
        "operationId": "CompaniesController_findOne_v1",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "Companies"
        ]
      },
      "patch": {
        "operationId": "CompaniesController_update_v1",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "Companies"
        ]
      },
      "delete": {
        "operationId": "CompaniesController_remove_v1",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "Companies"
        ]
      }
    },
...
    "/v1/action-plans": {
      "post": {
        "operationId": "ActionPlansController_create_v1",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateActionPlanDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "tags": [
          "Action Plans"
        ]
      },
      "get": {
        "operationId": "ActionPlansController_findAll_v1",
        "parameters": [
          {
            "name": "safetyReportId",
            "required": true,
            "in": "query",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "tags": [
          "Action Plans"
        ]
      }
    },
...
    "/v1/action-plans/{id}": {
      "get": {
        "operationId": "ActionPlansController_findOne_v1",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "tags": [
          "Action Plans"
        ]
      },
      "delete": {
        "operationId": "ActionPlansController_remove_v1",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "tags": [
          "Action Plans"
        ]
      }
    },
...
  },
  "info": {
    "title": "Hawana HSE Core API",
    "description": "HSE SaaS Core Backend (Multi-tenant)",
    "version": "1.0.0",
    "contact": {}
  },
...
  "servers": [],
  "components": {
    "securitySchemes": {
      "bearer": {
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "type": "http"
      }
    },
    "schemas": {
...
      "LoginDto": {
        "type": "object",
        "properties": {
          "email": {
            "type": "string",
            "example": "viewer@hawana.com"
          },
          "password": {
            "type": "string",
            "example": "Hawana@2026",
            "minLength": 6
          }
        },
        "required": [
          "email",
          "password"
        ]
      },
      "RefreshDto": {
        "type": "object",
        "properties": {
          "refreshToken": {
            "type": "string",
            "example": "tokenId.secret"
          }
        },
        "required": [
          "refreshToken"
        ]
      },
...
      "CreateActionPlanDto": {
        "type": "object",
        "properties": {
          "title": {
            "type": "string",
            "example": "Fix missing guardrail",
            "description": "Short title describing the corrective action"
          },
          "description": {
            "type": "string",
            "example": "Install temporary barrier until permanent fix is completed",
            "description": "Detailed description of the action plan"
          },
          "safetyReportId": {
            "type": "string",
            "example": "28925d55-83c4-4b73-a877-45c82d209528",
            "description": "Related Safety Report ID"
          },
          "assignedToUserId": {
            "type": "string",
            "example": "15fd5c5b-d594-4f08-b20c-0b1fb7262748",
            "description": "Assigned user ID"
          },
          "dueDate": {
            "type": "string",
            "example": "2026-03-01T00:00:00.000Z",
            "description": "Due date (ISO8601)"
          }
        },
        "required": [
          "title",
          "safetyReportId"
        ]
      },
...
    }
  }
}